Doc: Add documentation to the main.ts file
This commit adds documentation to
the main.ts file in the /src directory of
the jw-public repository. The documentation
consists of helper functions to check for
empty strings, check if a user is an admin,
and set initial values for error and success
messages. Additionally, the commit adds logic
to add or remove the "collapse" class in the
div with class "navbar-collapse" depending
on the width of the screen. This change will
ensure that the code is more readable and
easier to maintain going forward.